+++ src/builds/mewcx/cxHelpers/backgroundEvents/mew-cx-sign-msg.js
@@ -1,11 +1,12 @@
 /* eslint-disable no-undef */
 import { CX_SIGN_MSG, REJECT_MEW_SIGN_MSG } from '../cxEvents';
+import { toChecksumAddress } from '@/helpers/addressUtils';
 import { Misc } from '@/helpers';
 
 export default async ({ event, payload }, _, next) => {
   if (event !== CX_SIGN_MSG) return next();
-  chrome.storage.sync.get(payload.address, res => {
-    if (res[payload.address]) {
+  chrome.storage.sync.get(toChecksumAddress(payload.address), res => {
+    if (res[toChecksumAddress(payload.address)]) {
       chrome.windows.create({
         url: chrome.runtime.getURL(
           `index.html#/extension-popups/sign-msg?url=${payload.url}&msgToSign=${payload.msgToSign}&address=${payload.address}`