+++ src/layouts/InterfaceLayout/containers/SwapContainer/components/SwapExitToFiat/SwapExitToFiat.vue
@@ -28,8 +28,10 @@
                   :options="{
                     title: $t('swap.exit-to-fiat.iban-num'),
                     popover: $t('swap.exit-to-fiat.popover-iban-num'),
-                    value: ''
+                    value: '',
+                    validCheck: true
                   }"
+                  :valid-input="isValidIBAN"
                   @changedValue="orderDetails.iban = $event"
                 />
               </li>
@@ -80,9 +82,7 @@
                   <standard-input
                     :options="{
                       title: $t('swap.exit-to-fiat.billing-addr'),
-                      placeHolder: $t(
-                        'swap.exit-to-fiat.to-fiat.addr-placeholder'
-                      ),
+                      placeHolder: $t('swap.exit-to-fiat.addr-placeholder'),
                       value: ''
                     }"
                     class="address1"