+++ src/wallets/hybrid/MEWconnect/errorHandler.js
@@ -3,12 +3,12 @@ const ERRORS = {};
 const WARNING = {};
 
 export default err => {
-  const errorValues = Object.values(ERRORS);
-  const warningValues = Object.values(WARNING);
+  const errorValues = Object.keys(ERRORS);
+  const warningValues = Object.keys(WARNING);
   if (errorValues.includes(err.message)) {
-    Toast.responseHandler(err, Toast.ERROR);
+    Toast.responseHandler(ERRORS[err.message], Toast.ERROR);
   } else if (warningValues.includes(err.message)) {
-    Toast.responseHandler(err, Toast.WARN);
+    Toast.responseHandler(WARNING[err.message], Toast.WARN);
   } else {
     Toast.responseHandler(err, false);
   }