+++ vueConfig/mew.js
@@ -3,6 +3,7 @@ const UglifyJS = require('uglify-es');
 const env_vars = require('../ENV_VARS');
 const defaultConfigs = require('./defaultConfigs');
 const webpackConfig = {
+  devtool: defaultConfigs.devtool,
   node: {
     process: true
   },
@@ -15,9 +16,10 @@ const webpackConfig = {
   },
   plugins: defaultConfigs.plugins.concat([
     new CopyWebpackPlugin([
+      { from: 'security.txt', to: '.well-known/security.txt' },
       {
         from: 'src/builds/' + JSON.parse(env_vars.BUILD_TYPE) + '/public',
-        transform: function(content, filePath) {
+        transform: function (content, filePath) {
           if (filePath.split('.').pop() === ('js' || 'JS'))
             return UglifyJS.minify(content.toString()).code;
           return content;